requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-ollama
langchain-community
pydantic
pydantic-settings
python-dotenv
python-multipart
openai
redis
requests
unstructured
Jinja2
sentence-transformers
pypdf
python-docx
openpyxl
chromadb
